@@ -1,8 +1,9 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import warning from './routerWarning'
 import { routerShape } from './PropTypes'
 
-const { bool, object, string, func, oneOfType } = React.PropTypes
+const { bool, object, string, func, oneOfType } = PropTypes
 
 function isLeftClickEvent(event) {
   return event.button === 0
