@@ -1,12 +1,13 @@
 import invariant from 'invariant'
 import React from 'react'
+import PropTypes from 'prop-types'
 
 import deprecateObjectProperties from './deprecateObjectProperties'
 import getRouteParams from './getRouteParams'
 import { isReactChildren } from './RouteUtils'
 import warning from './routerWarning'
 
-const { array, func, object } = React.PropTypes
+const { array, func, object } = PropTypes
 
 /**
  * A <RouterContext> renders the component tree for a given router state
