@@ -1,5 +1,6 @@
 import expect from 'expect'
 import React, { cloneElement } from 'react'
+import PropTypes from 'prop-types'
 import { render } from 'react-dom'
 import Router from '../Router'
 import Route from '../Route'
@@ -11,8 +12,8 @@ const BAR_ROOT_CONTAINER_TEXT = 'BAR ROOT CONTAINER'
 const BAZ_CONTAINER_TEXT = 'BAZ INJECTED'
 
 const FooRootContainer = React.createClass({
-  propTypes: { children: React.PropTypes.node.isRequired },
-  childContextTypes: { foo: React.PropTypes.string },
+  propTypes: { children: PropTypes.node.isRequired },
+  childContextTypes: { foo: PropTypes.string },
   getChildContext() { return { foo: FOO_ROOT_CONTAINER_TEXT } },
   render() {
     return this.props.children
@@ -20,8 +21,8 @@ const FooRootContainer = React.createClass({
 })
 
 const FooContainer = React.createClass({
-  propTypes: { children: React.PropTypes.node.isRequired },
-  contextTypes: { foo: React.PropTypes.string.isRequired },
+  propTypes: { children: PropTypes.node.isRequired },
+  contextTypes: { foo: PropTypes.string.isRequired },
   render() {
     const { children, ...props } = this.props
     const fooFromContext = this.context.foo
@@ -39,8 +40,8 @@ const useFoo = () => ({
 })
 
 const BarRootContainer = React.createClass({
-  propTypes: { children: React.PropTypes.node.isRequired },
-  childContextTypes: { bar: React.PropTypes.string },
+  propTypes: { children: PropTypes.node.isRequired },
+  childContextTypes: { bar: PropTypes.string },
   getChildContext() { return { bar: BAR_ROOT_CONTAINER_TEXT } },
   render() {
     return this.props.children
@@ -48,8 +49,8 @@ const BarRootContainer = React.createClass({
 })
 
 const BarContainer = React.createClass({
-  propTypes: { children: React.PropTypes.node.isRequired },
-  contextTypes: { bar: React.PropTypes.string.isRequired },
+  propTypes: { children: PropTypes.node.isRequired },
+  contextTypes: { bar: PropTypes.string.isRequired },
   render() {
     const { children, ...props } = this.props
     const barFromContext = this.context.bar
